backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart==0.0.12
pydantic==2.10.4
annotated-types>=0.4.0,<0.7.0

# LangFlow
langflow==1.5.0.post1
langflow-base==0.5.0.post1
# Align with langflow's pinned dependency
chromadb==0.5.23

# LangChain ecosystem (required by provided flows)
langchain-core>=0.2.12
langchain-openai>=0.2.12
langchain-chroma>=0.1.0
langchain-text-splitters>=0.2.0

# Data & parsing
pypdf==5.1.0
pandas>=2.1.4
numpy>=1.26.4
plotly==5.17.0

# Auth & utils
aiofiles==24.1.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlmodel==0.0.22
openai>=1.0.0

# Tests
pytest==8.3.4 